Fix FFI attribute type: pass n as int32, not int64
With jax_enable_x64=True, Python int traces as int64 but the C++
FFI handler declares .Attr<int32_t>("n"). Cast to np.int32 before
passing to ffi_call.

Co-developed-by: Claude Code v2.1.109 (claude-opus-4-6)

+++ vajax/dense_cuda/dense_cuda_jax.py
@@ -15,6 +15,7 @@ from __future__ import annotations
 
 import jax
 import jax.extend.core
+import numpy as np
 from jax.core import ShapedArray
 from jax.interpreters import mlir
 from jaxtyping import Array
@@ -94,7 +95,7 @@ def _dense_lu_solve_f64_impl(A_flat: Array, neg_f: Array, *, n: int) -> Array:
         "dense_lu_solve_f64",
         jax.ShapeDtypeStruct((n,), neg_f.dtype),
     )
-    return call(A_flat, neg_f, n=n)
+    return call(A_flat, neg_f, n=np.int32(n))
 
 
 # =============================================================================
